restructure files
